+++ testing/test_se2.py
@@ -733,82 +733,6 @@ def test_ux_changing_pins(true_pin, repl, force_main_pin, goto_trick_menu,
 
     clear_all_tricks()
 
-def test_trick_backups(goto_trick_menu, clear_all_tricks, repl, unit_test, 
-        new_trick_pin, new_pin_confirmed, pick_menu_item, need_keypress):
-
-    clear_all_tricks()
-
-    # - make wallets of all duress types (x2 each)
-    # - plus a few simple ones
-    # - perform a backup and check result
-
-    for n in range(8):
-        goto_trick_menu()
-        pin = '123-%04d'%n
-        new_trick_pin(pin, 'Duress Wallet', None)
-        item = 'BIP-85 Wallet #%d' % (n%4) if (n%4 != 0) else 'Legacy Wallet'
-        pick_menu_item(item)
-        need_keypress('y')
-        new_pin_confirmed(pin, item, None, None)
-
-    for pin, op_mode, expect, _, xflags in [
-        ('11-33', 'Just Reboot', 'Reboot when this PIN', False, TC_REBOOT), 
-        ('11-55', 'Look Blank', 'Look and act like a freshly', False, TC_BLANK_WALLET), 
-    ]:
-        new_trick_pin(pin, op_mode, expect)
-        new_pin_confirmed(pin, op_mode, xflags)
-
-    # works, but not the best test
-    #unit_test('devtest/backups.py')
-
-    bk = repl.exec('import backups; RV.write(backups.render_backup_contents())', raw=1)
-
-    assert 'Coldcard backup file' in bk
-
-    def decode_backup(txt):
-        import json
-        vals = dict()
-        trimmed = dict()
-        for ln in txt.split('\n'):
-            if not ln: continue
-            if ln[0] == '#': continue
-
-            k,v = ln.split(' = ', 1)
-
-            v = json.loads(v)
-
-            if k.startswith('duress_') or k.startswith('fw_'):
-                # no space in USB xfer for thesE!
-                trimmed[k] = v
-            else:
-                vals[k] = v
-
-        return vals, trimmed
-
-    # decode it
-    vals, trimmed = decode_backup(bk)
-
-    assert 'duress_xprv' in trimmed
-    assert 'duress_1001_words' in trimmed
-    assert 'duress_1002_words' in trimmed
-    assert 'duress_1003_words' in trimmed
-
-    unit_test('devtest/clear_seed.py')
-    
-    repl.exec(f'import backups; backups.restore_from_dict_ll({vals!r})')
-
-    # recover from recovery
-    repl.exec(f'import backups; pa.setup(pa.pin); pa.login(); from actions import goto_top_menu; goto_top_menu()')
-
-    bk2 = repl.exec('import backups; RV.write(backups.render_backup_contents())', raw=1)
-    assert 'Traceback' not in bk2
-
-    vals2, tr2 = decode_backup(bk2)
-
-    assert vals == vals2
-    assert trimmed == tr2
-
-
 # TODO
 # - make trick and do login, check arrives right state?
 # - out of slots
