+++ testing/test_ux.py
@@ -117,157 +117,6 @@ def pass_word_quiz(need_keypress, cap_story):
 
     return doit
 
-@pytest.mark.qrcode
-@pytest.mark.parametrize('multisig', [False, 'multisig'])
-@pytest.mark.parametrize('st', ["b39pass", "eph", None])
-@pytest.mark.parametrize('reuse_pw', [False, True])
-@pytest.mark.parametrize('save_pw', [False, True])
-def test_make_backup(multisig, goto_home, pick_menu_item, cap_story, need_keypress, st,
-                     open_microsd, microsd_path, unit_test, cap_menu, word_menu_entry,
-                     pass_word_quiz, reset_seed_words, import_ms_wallet, get_setting,
-                     cap_screen_qr, reuse_pw, save_pw, settings_set, settings_remove,
-                     generate_ephemeral_words, set_bip39_pw, verify_backup_file,
-                     check_and_decrypt_backup, restore_backup_cs):
-    # Make an encrypted 7z backup, verify it, and even restore it!
-
-    # need to make multisig in my main wallet
-    if multisig and st != "eph":
-        import_ms_wallet(15, 15)
-        need_keypress('y')
-        time.sleep(.1)
-        assert len(get_setting('multisig')) == 1
-
-    if st == "b39pass":
-        xfp_pass = set_bip39_pw("coinkite", reset=False)
-        _, story = cap_story()
-        assert "Above is the master key fingerprint of the current wallet" in story
-        need_keypress("y")
-        assert not get_setting('multisig', None)
-    elif st == "eph":
-        eph_seed = generate_ephemeral_words(num_words=24, dice=False, from_main=True)
-        _, story = cap_story()
-        assert "New ephemeral master key in effect" in story
-        need_keypress("y")
-
-        if multisig:
-            # make multisig in ephemeral wallet
-            import_ms_wallet(15, 15, dev_key=True, common="605'/0'/0'")
-            need_keypress('y')
-            time.sleep(.1)
-            assert len(get_setting('multisig')) == 1
-
-    if reuse_pw:
-        settings_set('bkpw', ' '.join('zoo' for _ in range(12)))
-    else:
-        settings_remove('bkpw')
-
-    goto_home()
-    pick_menu_item('Advanced/Tools')
-    pick_menu_item('Backup')
-    pick_menu_item('Backup System')
-
-    title, body = cap_story()
-    if st:
-        if st == "b39pass":
-            assert "BIP39 passphrase is in effect" in body
-            assert "ignores passphrases and produces backup of main seed" in body
-            assert "(2) to back-up BIP39 passphrase wallet" in body
-        if st == "eph":
-            assert "An ephemeral seed is in effect" in body
-            assert "so backup will be of that seed" in body
-
-        need_keypress("y")
-        time.sleep(.1)
-        title, body = cap_story()
-
-    if reuse_pw:
-        assert ' 1: zoo' in body
-        assert '12: zoo' in body
-        need_keypress('y')
-        words = ['zoo']*12
-
-        time.sleep(0.1)
-        title, body = cap_story()
-    else:
-        assert title == 'NO-TITLE'
-        assert 'Record this' in body
-        assert 'password:' in body
-
-        words = [w[3:].strip() for w in body.split('\n') if w and w[2] == ':']
-        assert len(words) == 12
-
-        print("Passphrase: %s" % ' '.join(words))
-
-        if 'QR Code' in body:
-            need_keypress('1')
-            got_qr = cap_screen_qr().decode('ascii').lower().split()
-            assert [w[0:4] for w in words] == got_qr
-            need_keypress('y')
-
-        # pass the quiz!
-        count, title, body = pass_word_quiz(words)
-        assert count >= 4
-        assert "same words next time" in body
-        assert "Press (1) to save" in body
-        if save_pw:
-            need_keypress('1')
-            time.sleep(.1)
-
-            assert get_setting('bkpw') == ' '.join(words)
-        else:
-            need_keypress('x')
-            time.sleep(.01)
-            assert get_setting('bkpw', 'xxx') == 'xxx'
-
-        time.sleep(0.1)
-        title, body = cap_story()
-
-    time.sleep(0.1)
-    if st == "b39pass" and multisig:
-        # correct settings switch back?
-        # multisig is only in main wallet
-        # must not be copied from main to b39pass
-        # must not be available after backup done
-        assert not get_setting('multisig', None)
-
-    files = []
-    for copy in range(2):
-        if copy == 1:
-            title, body = cap_story()
-            assert 'written:' in body
-
-        fn = [ln.strip() for ln in body.split('\n') if ln.endswith('.7z')][0]
-
-        print("filename %d: %s" % (copy, fn))
-
-        files.append(fn)
-
-        # write extra copy.
-        need_keypress('2')
-        time.sleep(.01) 
-
-    bk_a = open_microsd(files[0]).read()
-    bk_b = open_microsd(files[1]).read()
-
-    assert bk_a == bk_b, "contents mismatch"
-
-    need_keypress('x')
-    time.sleep(.01)
-
-    verify_backup_file(fn)
-    check_and_decrypt_backup(fn, words)
-
-    for i in range(10):
-        need_keypress('x')
-        time.sleep(.01) 
-
-    # test verify on device (CRC check)
-    avail_settings = ['multisig'] if multisig else None
-    restore_backup_cs(files[0], words, avail_settings=avail_settings)
-
-    reset_seed_words()
-    settings_remove('multisig')
-
 
 @pytest.mark.qrcode
 @pytest.mark.parametrize('seed_words, xfp', [
